Add consistent task completion message for planned and normal tasks
