+++ src/Task.tsx
@@ -3,38 +3,7 @@ import PropTypes from 'prop-types';
 import TaskContent from './TaskContent'
 import React from 'react';
 import { FinishedState } from './TaskInterfaces.tsx';
-
-const translateLastCompletedDatetime = (lastCompletionDatetime) => {
-    let now = new Date()
-    let lastFinished = new Date(lastCompletionDatetime);
-
-    let timeDeltaInSeconds = Math.floor((now - lastFinished) / 1000);
-    let minutes = Math.floor(timeDeltaInSeconds / 60);
-    let hours = Math.floor(timeDeltaInSeconds / 3600);
-    let days = Math.floor(timeDeltaInSeconds / 86400);
-    let weeks = Math.floor(timeDeltaInSeconds / (86400 * 7))
-    let months = Math.floor(timeDeltaInSeconds / (86400 * 30.42))
-    let timeDeltaString = ""
-    if (timeDeltaInSeconds < 15) {
-        timeDeltaString = "just now"
-    } else if (timeDeltaInSeconds < 60) {
-        timeDeltaString = "less than a minute ago"
-    } else if (timeDeltaInSeconds < 3600) {
-        timeDeltaString = `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
-    } else if (timeDeltaInSeconds < 86400) {
-        timeDeltaString = `${hours} hour${hours > 1 ? 's' : ''} ago`;
-    } else if (days < 7) {
-        timeDeltaString = `${days} day${days > 1 ? 's' : ''} ago`;
-    } else if (weeks < 5) {
-        timeDeltaString = `${weeks} week${weeks > 1 ? 's' : ''} ago`;
-    } else {
-        timeDeltaString = `${months} month${months > 1 ? 's' : ''} ago`;
-    }
-
-
-
-    return timeDeltaString
-}
+import { translateLastCompletedDatetime } from './TaskHelperFuncs.js';
 
 const Task = (props) => {
     const { id, name, setTaskName, deleteTask,
@@ -258,9 +227,10 @@ const Task = (props) => {
     const textEditHandlers = { onChange: handleChange, onBlur: handleBlur, onKeyDown: handleKeyDown, onClick: captureClick(() => { }) }
     const selectHandlers = selected ? { onClick: captureClick(deleteFromSelection) } : { onClick: captureClick(addToSelection) }
     let fullName = name
-    if (repeated &&
+    if (
         (taskFinishStatus !== FinishedState.Impossible || taskFinishStatus !== FinishedState.Irrelevant) &&
-        taskLastCompletion
+        taskLastCompletion &&
+        !isEditing
     ) {
         fullName += " / " + translateLastCompletedDatetime(taskLastCompletion)
     }
